requirements.txt
# ===========================================
# Dependencias del Agente Calculadora
# ===========================================
# Instalar con: pip install -r requirements.txt

# -------------------------------------------
# LangChain - Framework para agentes
# -------------------------------------------
langchain>=0.1.0
langchain-openai>=0.0.5          # Para usar la API compatible con OpenAI (DeepSeek)
langchain-community>=0.0.20      # Herramientas de la comunidad (Wikipedia, etc.)
langgraph>=0.0.20                # Para agentes ReAct con LangGraph

# -------------------------------------------
# Herramientas de búsqueda
# -------------------------------------------
# SerpAPI (búsqueda web premium - opcional)
google-search-results>=2.4.2

# DuckDuckGo (búsqueda web gratuita - fallback)
duckduckgo-search>=4.1.0

# Wikipedia
wikipedia>=1.4.0

# -------------------------------------------
# Configuración
# -------------------------------------------
python-dotenv>=1.0.0             # Para cargar variables de .env

# -------------------------------------------
# Utilidades
# -------------------------------------------
requests>=2.31.0                  # HTTP requests
pydantic>=2.0.0    
uvicorn>=0.24.0               # Validación de datos
fastapi>=0.104.1
# -------------------------------------------
# Desarrollo (opcional)
# -------------------------------------------
# pytest>=7.4.0                   # Testing
# black>=23.0.0                   # Formateo de código
# mypy>=1.5.0                     # Type checking
